pyproject.toml
[project]
name = "bot-keuangan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"  # UBAH DARI >=3.14 KE >=3.10
dependencies = [
    "google-generativeai>=0.8.5",
    "groq>=0.37.1",
    "gspread>=6.2.1",
    "pandas>=1.5.0",
    "pillow>=10.2.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "pandasai==2.2.15",
    "seaborn>=0.12.0",
    "matplotlib>=3.7.0",
    "PyYAML>=6.0", # TAMBAHKAN INI
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "numpy<2.0.0",
    "pyngrok>=7.5.0",
]
